+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * **\[core\]** Allow `Observer` instances to be added to native component `Component` when the component has
   already been completed. This supports the use case where `when` observers are created when a new entity is
   added to a repository.
+* **\[component\]** Expose the native component in rhw `AbstractRepository` so that is available in extensions.
 
 ### [v0.63](https://github.com/arez/arez/tree/v0.63) (2018-03-07)
 [Full Changelog](https://github.com/arez/arez/compare/v0.62...v0.63)